Bug fix: Make sure 'git fetch' _always_ updates local branches
Detailed explanation: http://stackoverflow.com/a/10697486

To be honest I'm not sure why I never ran into this before, I've been
using this functionality for months and updates always came in just
fine based on the same version of git. Nevertheless the new 'git fetch'
command is the proper, documented way to do what I want git to do so
here we go :-).