Force internal merge tool for Mercurial*
After isolating the test suite from $HOME my ~/.hgrc
was ignored and the following setting disappeared:

[ui]
merge = internal:merge

Then I ran the vcs-repo-mgr test suite and `meld' popped up. Not what I
was expecting from an automated test suite! :-P

I think it's safe to say that this is a better default for the Mercurial
backend given the focus of vcs-repo-mgr on automation, if only to make
this backend match the behavior of the other backends.